+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 
 - "Delete All Whitespace at Point" now deletes whitespace to the left of caret after *Universal argument*.
+- "Transpose Lines" exchanges the current line with a line above it. Which line can be modified with *Universal argument*.
 
 ## [1.5.0] - 2025-09-30